+++ projects/igniteui-angular/src/lib/grids/grid/grid-filtering-ui.spec.ts
@@ -3258,21 +3258,24 @@ describe('IgxGrid - Filtering actions - Excel style filtering #grid', () => {
             grid.displayDensity = DisplayDensity.compact;
             tick(200);
             fix.detectChanges();
+            spyOn(grid.onColumnVisibilityChanged, 'emit');
 
             const column = grid.columns.find((col) => col.field === 'ProductName');
             GridFunctions.verifyColumnIsHidden(column, false, 6);
 
             // Open excel style filtering component and hide 'ProductName' column through header icon
-            GridFunctions.clickExcelFilterIcon(fix, 'ProductName');
-            tick(100);
-            fix.detectChanges();
+            GridFunctions.clickExcelFilterIconFromCode(fix, grid, 'ProductName');
             GridFunctions.clickHideIconInExcelStyleFiltering(fix);
             tick(200);
             fix.detectChanges();
 
             // Verify Excel menu is closed
             expect(GridFunctions.getExcelStyleFilteringComponent(fix)).toBeNull();
-            GridFunctions.verifyColumnIsHidden(column, true, 5);
+            expect(grid.onColumnVisibilityChanged.emit).toHaveBeenCalledTimes(1);
+            expect(grid.onColumnVisibilityChanged.emit).toHaveBeenCalledWith({
+                column: column,
+                newValue: true
+            });
         }));
 
         it('Should move pinned column correctly by using move buttons', fakeAsync(() => {
@@ -4240,6 +4243,7 @@ describe('IgxGrid - Filtering actions - Excel style filtering #grid', () => {
         }));
 
         it('Should filter grid through custom date filter dialog.', fakeAsync(() => {
+            const column = grid.getColumnByName('ReleaseDate');
             // Open excel style custom filtering dialog.
             GridFunctions.clickExcelFilterIcon(fix, 'ReleaseDate');
             tick(100);
@@ -4269,8 +4273,13 @@ describe('IgxGrid - Filtering actions - Excel style filtering #grid', () => {
             GridFunctions.clickApplyExcelStyleCustomFiltering(fix);
 
             // Verify the results are with 'today' date.
-            const cellValue = GridFunctions.getColumnCells(fix, 'ReleaseDate')[0].nativeElement;
-            expect(new Date(cellValue.innerText).toDateString()).toMatch(new Date().toDateString());
+            const pipe = new DatePipe(grid.locale);
+            const cell = GridFunctions.getColumnCells(fix, 'ReleaseDate')[0].nativeElement;
+            const filteredDate = SampleTestData.today;
+            const inputText = grid.datePipe.transform(filteredDate, column.pipeArgs.format, undefined, grid.locale);
+            const cellText = pipe.transform(filteredDate, column.pipeArgs.format, column.pipeArgs.timezone);
+            expect(cell.innerText).toMatch(cellText);
+            expect(datePickerInput.value).toMatch(inputText);
             expect(grid.filteredData.length).toEqual(1);
         }));
 
@@ -4281,6 +4290,7 @@ describe('IgxGrid - Filtering actions - Excel style filtering #grid', () => {
                 return newRec;
             });
             fix.detectChanges();
+            const column = grid.getColumnByName('ReleaseDate');
 
             // Open excel style custom filtering dialog.
             GridFunctions.clickExcelFilterIcon(fix, 'ReleaseDate');
@@ -4311,8 +4321,13 @@ describe('IgxGrid - Filtering actions - Excel style filtering #grid', () => {
             GridFunctions.clickApplyExcelStyleCustomFiltering(fix);
 
             // Verify the results are with 'today' date.
-            const cellValue = GridFunctions.getColumnCells(fix, 'ReleaseDate')[0].nativeElement;
-            expect(new Date(cellValue.innerText).toDateString()).toMatch(new Date().toDateString());
+            const pipe = new DatePipe(grid.locale);
+            const cell = GridFunctions.getColumnCells(fix, 'ReleaseDate')[0].nativeElement;
+            const filteredDate = SampleTestData.today;
+            const inputText = grid.datePipe.transform(filteredDate, column.pipeArgs.format, undefined, grid.locale);
+            const cellText = pipe.transform(filteredDate, column.pipeArgs.format, column.pipeArgs.timezone);
+            expect(cell.innerText).toMatch(cellText);
+            expect(datePickerInput.value).toMatch(inputText);
             expect(grid.filteredData.length).toEqual(1);
         }));
 
@@ -4321,6 +4336,126 @@ describe('IgxGrid - Filtering actions - Excel style filtering #grid', () => {
                 const newRec = Object.assign({}, rec) as any;
                 newRec.ReleaseDate = rec.ReleaseDate ? rec.ReleaseDate.getTime() : null;
                 return newRec;
+            });
+            fix.detectChanges();
+            const column = grid.getColumnByName('ReleaseDate');
+
+            // Open excel style custom filtering dialog.
+            GridFunctions.clickExcelFilterIcon(fix, 'ReleaseDate');
+            tick(100);
+            fix.detectChanges();
+            GridFunctions.clickExcelFilterCascadeButton(fix);
+            fix.detectChanges();
+            GridFunctions.clickOperatorFromCascadeMenu(fix, 0);
+            tick(200);
+
+            const expr = GridFunctions.getExcelCustomFilteringDateExpressions(fix)[0];
+            const datePicker = expr.querySelector('igx-date-picker');
+            const datePickerInput = datePicker.querySelector('input');
+
+            // Click date picker input to open calendar.
+            datePickerInput.dispatchEvent(new MouseEvent('click'));
+            tick(100);
+            fix.detectChanges();
+
+            // Click today item.
+            const calendar = document.querySelector('igx-calendar');
+            const todayItem = calendar.querySelector('.igx-calendar__date--current');
+            (todayItem as HTMLElement).click();
+            tick(100);
+            fix.detectChanges();
+
+            // Click 'apply' button to apply filter.
+            GridFunctions.clickApplyExcelStyleCustomFiltering(fix);
+
+            // Verify the results are with 'today' date.
+            const pipe = new DatePipe(grid.locale);
+            const cell = GridFunctions.getColumnCells(fix, 'ReleaseDate')[0].nativeElement;
+            const filteredDate = SampleTestData.today;
+            const inputText = grid.datePipe.transform(filteredDate, column.pipeArgs.format, undefined, grid.locale);
+            const cellText = pipe.transform(filteredDate, column.pipeArgs.format, column.pipeArgs.timezone);
+            expect(cell.innerText).toMatch(cellText);
+            expect(datePickerInput.value).toMatch(inputText);
+            expect(grid.filteredData.length).toEqual(1);
+        }));
+
+        it('Should filter grid through custom date filter dialog when using pipeArgs for the column', fakeAsync(() => {
+            fix.componentInstance.data = SampleTestData.excelFilteringData().map(rec => {
+                const newRec = Object.assign({}, rec) as any;
+                newRec.ReleaseDate = rec.ReleaseDate ? rec.ReleaseDate.toISOString() : null;
+                return newRec;
+            });
+
+            const formatOptions = {
+                timezone: 'utc',
+            };
+            const column = grid.getColumnByName('ReleaseDate');
+            column.pipeArgs = formatOptions;
+            fix.detectChanges();
+
+            // Open excel style custom filtering dialog.
+            GridFunctions.clickExcelFilterIcon(fix, 'ReleaseDate');
+            tick(100);
+            fix.detectChanges();
+            GridFunctions.clickExcelFilterCascadeButton(fix);
+            fix.detectChanges();
+            GridFunctions.clickOperatorFromCascadeMenu(fix, 0);
+            tick(200);
+
+            const expr = GridFunctions.getExcelCustomFilteringDateExpressions(fix)[0];
+            const datePicker = expr.querySelector('igx-date-picker');
+            const datePickerInput = datePicker.querySelector('input');
+
+            // Click date picker input to open calendar.
+            datePickerInput.dispatchEvent(new MouseEvent('click'));
+            tick(100);
+            fix.detectChanges();
+
+            // Click today item.
+            const calendar = document.querySelector('igx-calendar');
+            const days = calendar.querySelectorAll('igx-day-item');
+            const todayItem = calendar.querySelector('.igx-calendar__date--current');
+            let todayIndex;
+            let i = 0;
+            while (!todayIndex) {
+                const item = days[i];
+                if (item === todayItem) {
+                    todayIndex = i;
+                }
+                i++;
+            }
+
+            // find the day 15 days later than today
+            const day = days[todayIndex + 15];
+            (day as HTMLElement).click();
+            tick(100);
+            fix.detectChanges();
+
+            // Click 'apply' button to apply filter.
+            GridFunctions.clickApplyExcelStyleCustomFiltering(fix);
+
+            // Verify the results. Filtered day is date that is 15 days after today
+            const pipe = new DatePipe(grid.locale);
+            const cell = GridFunctions.getColumnCells(fix, 'ReleaseDate')[0].nativeElement;
+            const filteredDate = SampleTestData.timeGenerator.timedelta(SampleTestData.today, 'day', 15);
+            const inputText = grid.datePipe.transform(filteredDate, column.pipeArgs.format, undefined, grid.locale);
+            const cellText = pipe.transform(filteredDate, column.pipeArgs.format, column.pipeArgs.timezone);
+            expect(cell.innerText).toMatch(cellText);
+            expect(datePickerInput.value).toMatch(inputText);
+            expect(grid.filteredData.length).toEqual(1);
+        }));
+
+        it('Should filter grid through custom date filter dialog when using pipeArgs and formatter for the column', fakeAsync(() => {
+            registerLocaleData(localeFR);
+            const pipe = new DatePipe('fr-FR');
+            const formatOptions = {
+                timezone: 'utc',
+            };
+            const column = grid.getColumnByName('ReleaseDate');
+            column.pipeArgs = formatOptions;
+            grid.getColumnByName('ReleaseDate').formatter = ((value: any) => {
+                const val = value !== null && value !== undefined && value !== '' ? pipe.transform(value, 'longDate') : 'No value!';
+                return val;
             });
             fix.detectChanges();
 
@@ -4344,17 +4479,34 @@ describe('IgxGrid - Filtering actions - Excel style filtering #grid', () => {
 
             // Click today item.
             const calendar = document.querySelector('igx-calendar');
+            const days = calendar.querySelectorAll('igx-day-item');
             const todayItem = calendar.querySelector('.igx-calendar__date--current');
-            (todayItem as HTMLElement).click();
+            let todayIndex;
+            let i = 0;
+            while (!todayIndex) {
+                const item = days[i];
+                if (item === todayItem) {
+                    todayIndex = i;
+                }
+                i++;
+            }
+
+            // find the day 15 days later than today
+            const day = days[todayIndex + 15];
+            (day as HTMLElement).click();
             tick(100);
             fix.detectChanges();
 
             // Click 'apply' button to apply filter.
             GridFunctions.clickApplyExcelStyleCustomFiltering(fix);
 
-            // Verify the results are with 'today' date.
-            const cellValue = GridFunctions.getColumnCells(fix, 'ReleaseDate')[0].nativeElement;
-            expect(new Date(cellValue.innerText).toDateString()).toMatch(new Date().toDateString());
+            // Verify the results. Filtered day is date that is 15 days after today
+            const cell = GridFunctions.getColumnCells(fix, 'ReleaseDate')[0].nativeElement;
+            const filteredDate = SampleTestData.timeGenerator.timedelta(SampleTestData.today, 'day', 15);
+            const inputText = column.formatter(filteredDate);
+            const cellText = column.formatter(filteredDate);
+            expect(cell.innerText).toMatch(cellText);
+            expect(datePickerInput.value).toMatch(inputText);
             expect(grid.filteredData.length).toEqual(1);
         }));
 
@@ -5380,6 +5532,41 @@ describe('IgxGrid - Filtering actions - Excel style filtering #grid', () => {
 
             const loadingIndicator = GridFunctions.getExcelFilteringLoadingIndicator(fix);
             expect(loadingIndicator).toBeNull('esf loading indicator is visible');
+        }));
+
+        it('Verify ESF list items when different date values produce same value label in the ESF list', fakeAsync(() => {
+            const grid = fix.componentInstance.grid;
+            grid.getColumnByName('ReleaseDate').formatter = ((value: any) => {
+                const pipe = new DatePipe(grid.locale);
+                const val = value !== null && value !== undefined && value !== '' ? pipe.transform(value, 'longTime') : 'No value!';
+                return val;
+            });
+            fix.detectChanges();
+
+            // Open excel style custom filtering dialog and wait a bit.
+            GridFunctions.clickExcelFilterIcon(fix, 'ReleaseDate');
+            tick(1050);
+            fix.detectChanges();
+
+            // Verify items in search have loaded and that the loading indicator is not visible.
+            const listItems = GridFunctions.getExcelStyleSearchComponentListItems(fix);
+            expect(listItems.length).toBe(3, 'incorrect rendered list items count');
+
+            const checkboxElements = GridFunctions.getExcelStyleFilteringCheckboxes(fix);
+            checkboxElements[0].click();
+            checkboxElements[2].click();
+            tick();
+            fix.detectChanges();
+
+            GridFunctions.clickApplyExcelStyleFiltering(fix);
+            fix.detectChanges();
+
+            expect(grid.filteredData.length).toEqual(1);
+            expect(grid.filteredData[0].Downloads).toEqual(254);
+
+            tick(1050);
+            fix.detectChanges();
+
         }));
 
         it('Done callback should be executed only once per column', fakeAsync(() => {